backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
langchain>=0.3.27
langchain-community>=0.3.27
langchain-core>=0.3.74
langchain-google-genai>=2.1.9
langchain-openai>=0.3.30
langchain-postgres>=0.0.15
langchain-text-splitters>=0.3.9
psycopg[binary]>=3.2.9
pgvector>=0.3.6
pypdf>=6.0.0
python-dotenv>=1.1.1
pydantic-settings>=2.10.1
